fix: 클라이언트, 서버 컴포넌트 분리 (fail)
